[GraphScheduler] Split scheduler into .h and .cpp
Description:
This commit splits the code in GraphScheduler.cpp into separate
header and source files. This will allow scheduler unit tests to be
added in future commits.

Testing:
All existing unit tests pass.
